fix : delete Login.tsx & toggle Switch Renewal
