+++ src/sass/abstracts/_mixins.scss
@@ -1,0 +1,23 @@
+/*반응형 화면 크기 변수*/
+$mobile: 480px;
+$tablet_min: 481px;
+$tablet_max: 1079px;
+$desktop: 1080px;
+
+@mixin mobile {
+  @media (max-width: $mobile) {
+    @content;
+  }
+}
+
+@mixin tablet {
+  @media ((min-width: $tablet_min) and (max-width: $tablet_max)) {
+    @content;
+  }
+}
+
+@mixin desktop {
+  @media (min-width: $desktop) {
+    @content;
+  }
+}
