+++ src/sass/abstracts/_variables.scss
@@ -1,0 +1,3 @@
+// main-color
+$main-color-1: #0e0171;
+$main-color-2: #0cdcbf;
